requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
Flask==2.2.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.41.1
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.15.0
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.7.2
mediapipe==0.10.2
ml-dtypes==0.2.0
numpy==1.26.3
oauthlib==3.2.2
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.2
Pillow==9.5.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sounddevice==0.4.6
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==2.3.0
wrapt==1.14.1
zipp==3.17.0
